Fixed the bug where you add money to player crashes the app due to decimal balance

+++ frontend/AndroidCasinoUser/app/src/main/java/com/davidzhang/androidcasinouser/CreateAccount.java
@@ -81,7 +81,7 @@ public class CreateAccount extends ThemedActivity {
                         currentUser.setId(user.getString("_id"));
                         currentUser.setUserId(user.getString("userId"));
                         currentUser.setUsername(user.getString("username"));
-                        currentUser.setBalance(user.getInt("balance"));
+                        currentUser.setBalance(user.getDouble("balance"));
                         currentUser.setAdmin(user.getBoolean("isAdmin"));
                         currentUser.setChatBanned(user.getBoolean("isChatBanned"));
                         currentUser.setLastRedemptionDate(user.getString("lastRedemptionDate"));
